Refactored the API types to the index
