+++ packages/lytenyte-core/src/types/index.ts
@@ -1,7 +1,13 @@
-export type * from "./api.js";
+import type { Grid } from "../index.js";
+
 export type * from "./events.js";
 export type * from "./column.js";
 export type * from "./grid.js";
+
+export type Props<Spec extends Grid.GridSpec = Grid.GridSpec> = Grid.Props<Spec>;
+export type API<Spec extends Grid.GridSpec = Grid.GridSpec> = Grid.API<Spec>;
+export type ExportDataRectResult<Spec extends Grid.GridSpec = Grid.GridSpec> =
+  Grid.T.ExportDataRectResult<Spec>;
 
 export type { ReactPlaceholderFn, DragItem, DragItemTransfer, DragItemSiteLocal } from "../dnd/types.js";
